fix(Core): 优化了`Permission`/`Router`/`Config`代码
